Merge:
- Resolve ack.h conflict: keep both dlg_state/dlg_pending fields and
  the new REVENANT_DATA* field added by main.

Spec completion:
- Implement npc_dialogue_sanitize_input(): strip role-boundary injection
  tokens ([INST], <<SYS>>, <|system|>, </s>, <|im_start|>, <|im_end|>)
  and replace angle brackets with parentheses before dispatch.
- Add injection keyword short-circuit in npc_dialogue_dispatch(): detect
  "ignore previous", "you are now", "act as", etc. and enqueue a fixed
  confused refusal without calling the API; log attempt at log_f level.
- Extract max_tokens:100 to TNGAI_MAX_TOKENS constant in config.h.
- Rename all OPENCLAW_* constants to TNGAI_* (URL, MODEL, TIMEOUT,
  MAX_TOKENS) and update call_openclaw() to call_tngai() throughout,
  since OpenClaw is no longer used.

Documentation:
- Add docs/npc_dialogue_api.md with full call-flow ASCII diagram,
  tng-ai request/response reference, configuration constants table,
  threading model diagram, system prompt assembly layers, and prompt
  injection defence details.

+++ src/headers/strfuns.h
@@ -68,7 +68,7 @@ char *affect_loc_name args((int location));
 char *affect_bit_name args((int vector));
 char *raffect_bit_name args((int vector));
 
-char *extra_bit_name args((int extra_flags));
+char *extra_bit_name args((unsigned long long extra_flags));
 char *race_name args((CHAR_DATA * ch));
 char *short_race_name args((CHAR_DATA * ch));
 char *get_adept_name args((CHAR_DATA * ch));
